FEM: unit tests, move helpers in separate test tools module
