+++ src/Mod/Fem/CMakeLists.txt
@@ -116,6 +116,7 @@ SET(FemSolverFenics_SRCS
 SET(FemTests_SRCS
     femtest/__init__.py
     femtest/testfemcommon.py
+    femtest/testtools.py
 )
 
 SET(FemTestsFiles_SRCS
